+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Busy indicator when the list of entries is empty for a more responsive look
+
 ## [0.2] - 2016-11-21
 ### Added
 - System tray icon